# internal/bytealg: optimize Index/IndexString in amd64